Добавил шаблон для подтверждения удаления поста

+++ blog/views.py
@@ -64,6 +64,9 @@ def update_post(request, post_id):
 def delete_post(request, post_id):
     post = get_object_or_404(Post, id=post_id)
 
-    post.delete()
+    if request.method == "POST":
+        post.delete()
 
-    return redirect("post_list")
+        return redirect("post_list")
+    
+    return render(request, 'blog/confirm_post_delete.html', {'post': post})
